+++ tests/processing/pre/test_renaming.py
@@ -1,11 +1,18 @@
-from app.processing.graph import ProgramNode, QasmImplementation, SectionInfo
+from app.processing.graph import (
+    ProgramNode,
+    QasmImplementation,
+    SectionInfo,
+    SnippetIOInfo,
+)
 from app.processing.pre.renaming import RenameRegisterTransformer
 from tests.processing.utils import assert_processor
 
 
 def test_register_renaming() -> None:
     section_info = SectionInfo(
-        1, node=ProgramNode("42", QasmImplementation.create("qubit a;"), None)
+        1,
+        node=ProgramNode("42", QasmImplementation.create("qubit a;"), None),
+        io=SnippetIOInfo(),
     )
     assert_processor(
         RenameRegisterTransformer(),
